requirements.txt
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.35.0
flask>=2.3.0
nltk>=3.8.1
spacy>=3.7.0
textblob>=0.17.1
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
tiktoken>=0.5.0
